Eliminate all 5 remaining global locks: lock-free kernel objects
Two-track approach to remove every global SpinLock from the kernel:

Track 1 — Port-referenced kernel objects:
- ASPACE_TABLE → port-based resolution + per-aspace SpinLock (ASpaceId = u64)
- OBJ_TABLE → port-based resolution + per-object SpinLock (ObjectId = u64)
  COW sharing via atomic frame refcounts (O(1)) instead of O(N) sibling scan
- CAP_SYSTEM → per-task cap_lock + narrow CDT_LOCK
  Grant/revoke on different tasks now runs in parallel

Track 2 — Allocator scalability:
- FRAME_TABLE → AtomicU32 per-page refcounts (lock-free inc/dec/get)
- Slab caches → per-CPU magazine pairs (32-entry loaded+backup)
  Fast path: IRQs-disabled pop/push, no lock. Lock only on magazine swap.

Supporting changes:
- CapSet dead-port eviction when 64 inline slots fill up
- Slab magazine CPU migration fix (re-read cpu_id after IRQ restore)
- ExtentEntry.object_id widened to u64 (LEAF_CAP 17→15 for node fit)
- Type cascading: ObjectId/ASpaceId u32→u64 across mm/, sched/, syscall/

+++ kernel/src/mm/fault.rs
@@ -330,7 +330,7 @@ fn try_contiguous_promotion(pt_root: usize, vma: &super::vma::Vma, mmu_idx: usiz
 fn try_superpage_promotion(
     pt_root: usize,
     vma: &mut super::vma::Vma,
-    obj_id: u32,
+    obj_id: u64,
     mmu_idx: usize,
 ) {
     let mmu_count = vma.mmu_page_count();
@@ -447,7 +447,7 @@ fn try_superpage_promotion(
 pub fn try_superpage_promotion_eager(
     pt_root: usize,
     vma: &mut super::vma::Vma,
-    obj_id: u32,
+    obj_id: u64,
 ) {
     let mmu_count = vma.mmu_page_count();
     if mmu_count < SUPER_MMU_PAGES {
@@ -569,7 +569,7 @@ pub fn demote_superpage(pt_root: usize, va: usize, flags: u64) -> bool {
 fn handle_cow_fault(
     pt_root: usize,
     vma: &mut super::vma::Vma,
-    obj_id: u32,
+    obj_id: u64,
     obj_page_idx: usize,
     mmu_idx: usize,
     fault_addr: usize,
